pyproject.toml
[tool.poetry]
name = "youtube-to-whisper-gpu"
version = "0.1.0"
description = ""
authors = ["Sean Dearnaley <seandearnaley@hotmail.com>"]
readme = "README.md"
packages = [{include = "youtube_to_whisper_gpu"}]

[tool.poetry.dependencies]
python = "^3.10"
openai-whisper = "^20230314"
pytube = "^12.1.3"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"
ruff = "^0.0.261"
mypy = "^1.2.0"
black = "^23.3.0"
pytest = "^7.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
